# Perform some optimizations during interpreter code generation