This PR implements two optimizations for the interpreter.

1. When generating `MCode`, we look up arities for functions and emit more efficient `Call` instructions. This allows things like self-recursive loops to be more efficient.
    - It also fixes an opposite-logic bug with the stack size check in that path. Chris noticed it a while back, but it hadn't mattered because it was never used.
2. Just before generating `MCode`, it runs an inlining pass on the A-normal code. This checks for really simple wrapper-like code like (conceptually) `x y -> f x y y`, or stuff like `x -> 3`. If such things are fully or over-saturated, it will try to inline directly into the call site. Eventually, I think most builtins will be of this form, so saturated calls to builtins will be inlined to directly use the interpreter instructions.
   - Like I said, this pass happens _just_ before generating `MCode`, so it only affects what is actually run on the interpreter. We can't really do it earlier, because the inlined code isn't proper to send around to other machines, and the runtime hashes would change.

Also included are some tweaks to the pretty printing for debugging `MCode` generation. I think I'm going to try one more tweak to that before merging this.