@dolio 2024-10-29
Results of `@pchiusano/misc-benchmarks`

Without optimizations:

```
fib1
1.620115ms

fib2
5.263945ms

fib3
6.05794ms

Decode Nat
1.107µs

Generate 100 random numbers
750.962µs

List.foldLeft
5.181553ms

Count to 1 million
586.405189ms

Json parsing (per document)
657.471µs

Count to N (per element)
769ns

Count to 1000
768.651µs

Mutate a Ref 1000 times
1.3586ms

CAS an IO.ref 1000 times
1.799077ms

List.range (per element)
1.047µs

List.range 0 1000
1.110156ms

Set.fromList (range 0 1000)
5.253066ms

Map.fromList (range 0 1000)
3.9308ms

NatMap.fromList (range 0 1000)
18.409758ms

Map.lookup (1k element map)
11.931µs

Map.insert (1k element map)
23.333µs

List.at (1k element list)
1.25µs

Text.split /
72.824µs
```

With optimizations:

```
fib1
1.4078ms

fib2
4.922104ms

fib3
5.670696ms

Decode Nat
932ns

Generate 100 random numbers
647.73µs

List.foldLeft
4.492801ms

Count to 1 million
514.030489ms

Json parsing (per document)
611.919µs

Count to N (per element)
651ns

Count to 1000
668.335µs

Mutate a Ref 1000 times
1.114969ms

CAS an IO.ref 1000 times
1.491692ms

List.range (per element)
887ns

List.range 0 1000
956.614µs

Set.fromList (range 0 1000)
4.234707ms

Map.fromList (range 0 1000)
3.200286ms

NatMap.fromList (range 0 1000)
16.115533ms

Map.lookup (1k element map)
7.848µs

Map.insert (1k element map)
20.428µs

List.at (1k element list)
822ns

Text.split /
69.549µs
```